requirements.txt
pycurl>=7.21.5
progressbar2
